ui: extract OwnPool component
