+++ app/src/views/PoolPage/PoolPageV2.tsx
@@ -1,9 +1,6 @@
-import { useRowanPrice } from "@/hooks/useRowanPrice";
-import { ElementOf } from "@/utils/types";
-import { computed, defineComponent, PropType, ref } from "vue";
+import { computed, defineComponent } from "vue";
 import { RouterView } from "vue-router";
 
-import AssetPair from "~/components/AssetPair";
 import { Button } from "~/components/Button/Button";
 import Layout from "~/components/Layout";
 import PageCard from "~/components/PageCard";
@@ -11,20 +8,18 @@ import { SearchBox } from "~/components/SearchBox";
 import { useRemoveLiquidityMutation } from "~/domains/clp/mutation/liquidity";
 import { useCurrentRewardPeriod } from "~/domains/clp/queries/params";
 import { flagsStore, isAssetFlaggedDisabled } from "~/store/modules/flags";
+
 import {
   CompetitionsBySymbolLookup,
   useLeaderboardCompetitions,
 } from "../LeaderboardPage/useCompetitionData";
+import OwnPool from "./children/OwnPool";
 import {
+  PoolDataArray,
   PoolPageColumnId,
   PoolRewardProgram,
   usePoolPageData,
 } from "./usePoolPageData";
-import { useUserPoolData } from "./useUserPoolData";
-
-type PoolDataArray = ReturnType<
-  typeof usePoolPageData
->["allPoolsData"]["value"];
 
 export default defineComponent({
   name: "PoolsPage",
@@ -216,55 +211,6 @@ export default defineComponent({
           )} */}
         </PageCard>
       </Layout>
-    );
-  },
-});
-
-const OwnPool = defineComponent({
-  props: {
-    context: {
-      type: Object as PropType<ElementOf<PoolDataArray>>,
-      required: true,
-    },
-  },
-
-  setup(props) {
-    const userPoolData = useUserPoolData({
-      externalAsset: computed(() => props.context.pool.externalAmount.symbol),
-    });
-    const rowanPrice = useRowanPrice();
-    const { pool, accountPool } = props.context;
-
-    const details = [
-      ["Your pool share %", `${userPoolData.myPoolShare.value ?? "0"}%`],
-      ["Your pool value", `${userPoolData.myPoolShare.value ?? "0"}%`],
-      ["Rewards earned", `${userPoolData.myPoolShare.value ?? "0"}%`],
-      ["Performance", `${userPoolData.myPoolShare.value ?? "0"}%`],
-    ];
-    return () => (
-      <li
-        role="article"
-        class="flex h-[312px] flex-col gap-4 rounded-lg bg-gray-900 p-4"
-      >
-        <header class="flex items-center p-2">
-          <AssetPair asset={ref(pool.externalAmount.asset)} invert />
-        </header>
-        <main class="flex-1">
-          <ul class="grid w-full gap-2">
-            {details.map(([label, value]) => (
-              <li class="flex items-center justify-between">
-                <div class="text-gray-sif_200 text-left">{label}</div>
-                <div class="text-right">{value}</div>
-              </li>
-            ))}
-          </ul>
-        </main>
-        <footer class="">
-          {pool.externalAmount.displaySymbol.startsWith("a") && (
-            <Button.CallToAction>Claim</Button.CallToAction>
-          )}
-        </footer>
-      </li>
     );
   },
 });
